fixed minor typos and raising attack
attack now goes up by .5 and takes double instead of int

also deleted class files

+++ Character.java
@@ -17,7 +17,7 @@ public abstract class Character {
 	return _hp > 0;
     }
 
-    public void raiseAttack(int x) {
+    public void raiseAttack(double x) {
 	_attack += x;
     }
